requirements.txt
PyYAML>=6.0.1
GitPython>=3.1.40
radon>=6.0.1
matplotlib>=3.8.2
plotly>=5.18.0
Jinja2>=3.1.3
requests>=2.31.0
pygments>=2.17.2
pandas>=2.1.4
lizard>=1.17.10
